fix test lazy and, or. add  wildcards test
