+++ Functional/src/test/java/ru/spbau/mit/Wildcards.java
@@ -1,0 +1,14 @@
+package ru.spbau.mit;
+
+public class Wildcards {
+    static class Base {
+
+    }
+    static class Derived extends Base {
+
+    }
+    static class DerivedDerived extends Derived {
+
+    }
+
+}
